BLU: Improvements to how we handle weaving & Surpanakhas
Surpanakha is an oGCD with 4 charges.  Each time you use it,
it gives you a stacking buff that makes the next Surpanakha stronger,
but only IF the next action is a Surpanakha.

Doing absolutely anything other than a Surpanakha while the buff
is up will completely drop it.

Weaving before or after the Surpanakha spam will always result in
fully clipping the following GCD.

So far, this is what the existing Weaving module for BLU did; it
checked if you had a Surpanakha x4 weave and gave you a pass.

However, a problem: the standard BLU opener actually subverts this!
It hexaweaves the Surpanakhas to make them fit inside of a Moon Flute.

So to avoid some useless noise, these changes allows for a
penta/hexaweave during the Moon Flute window IF the Surpanakhas are
at the end of the weave.

And since I was already tweaking this, it also adds new suggestion
if Surpanakha is used in sets of less than 4, or if other abilities
are used in-between the sequence, dropping the buff.