BLU: Track lost Phantom Flurry ticks
Phantom Flurry is a channel that deals a 200 potency hit for every
second that it is channeled.

DPSes generally want to channel the entire 5 second duration at the
end of the Moon Flute window, so losing even one or two ticks
is a fair bit of potency.

Non-DPSes on the other hand want to get as close as possible to 5
seconds, then press the button again for an extra 600 potency hit.

This is adding suggestions for both; one for missing ticks, and
one for not pressing the button again if outside of a moon flute.