BLU: Track a bunch of useful actions in the timeline
Note that this is somewhat buggy at the moment due to abilities
with shared cooldowns, like Matra Magic / Dragon Force, or
Eruption / Feather Rain.

Currently the timeline code just picks whatever was listed first
in the code; I've spent a while trying to make it display whatever
was first used, but we don't have that data when setting up
the timeline rows, and delaying setting those up breaks things
in ways I don't quite understand.

Regardless, this is groundwork to build on.