BLU Action Timeline: Minor bandaid for Quasar/J Kick
These two share a CD and so should be grouped together, but having
the action timeline core module show a different header depending
on which one was used has been giving me trouble.

So this is just a quick bandaid for Quasar / J Kick.  The problem is
still there for Angel's Snack / Dragon Force / Matra Magic.