Core ActionTimeline: Option to resolve the label name late
This is really just a nice QoL for BLU; we have several abilities
that share cooldowns, like Quasar and J Kick -- people will only take
one or the other, so it's much nicer if the timeline shows exactly
the name they want.

Internally, this introduces a `lateResolveLabel` attribute to
the action row config; if provided and true, then when passing
multiple alternative abilities for one row, it will wait until
the last possible moment, then look at the cooldowns and find which
one of the mutually exclusive abilities was used.